+++ src/main/resources/db/changelog/changes/003-create-carts-table.sql
@@ -1,6 +1,6 @@
 create table carts(
     id bigint not null primary key,
-    total double precision,
+    total decimal(10, 2),
     status varchar(255),
     user_id bigint,
     constraint FK_USER_CART foreign key (user_id) references users (id)
